[Chore] Add implementation for FCM / Refactor Notification Class
