+++ app/src/main/java/com/angdroid/refrigerator_manament/application/Notification.kt
@@ -1,4 +1,4 @@
-package com.angdroid.refrigerator_manament.presentation.custom
+package com.angdroid.refrigerator_manament.application
 
 import android.app.NotificationChannel
 import android.app.NotificationManager
@@ -9,7 +9,6 @@ import android.content.Intent
 import android.os.Build
 import androidx.core.app.NotificationCompat
 import com.angdroid.refrigerator_manament.R
-import com.angdroid.refrigerator_manament.presentation.SplashActivity
 import com.angdroid.refrigerator_manament.presentation.home.HomeActivity
 
 class Notification(private val context: Context) {
